move oai and templates
